ShowConfig: More compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@40485 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/tools/ShowConfig.c
@@ -151,7 +151,7 @@ static VOID PrintProcessorInformation()
 
         printf("PROCESSOR %d:\t[%s/%s] %s", (int)(i + 1), architecturestring, endiannessstring, modelstring);
         if (cpuspeed)
-            printf(" (%llu Mhz)", cpuspeed / 1000000);
+            printf(" (%llu Mhz)", (unsigned long long)(cpuspeed / 1000000));
         printf("\n");
     }
 }
